mod: prefer raw pointer use when referencing object
When an object is passed ot another class or function, the need of smart pointers
is overkill and can be confusing when shared or unique ownership is required.
Raw pointers are in this case use to symbolize a reference (that can be null)
to an object. This object can be used by said class *without* gaining ownership
which is the desired behavior.

TL/DR: The use of raw pointers clarifies the intention of use of an object