fix: cpu load delays (#199)
* fix: load delay when successive loads target same register

When 2 load instructions use the same register and follow each other we have
to reset the load delay to this register because the second load instruction
overrides the first load instruction.

* fix: only override pending load in loadWithDelay function

The LWL and LWR instructions have to check if a pending load is waiting to
be processed. If there is a pending load, these instructions need to take the
pending value as the "register" value onto which they will operate.

In the previous implementation, LWR and LWL checked if there was a pending
load in the queue but they did not check if they were targeting the same
register as the pending load, which lead to issues when overriding pending
loads.