fix: set return address before jumping with jalr
When using the same register for saving the return address as well
as for the jump destination address, jalr was first setting the
return address into the register which was overriding the jump
destination address, effectively canceling the jump.

Now jalr first sets the jump destination address before saving the
return address into the register.